Hmmmm...I have no idea what to do at this point. The update is just so crazy. I am shocked the network can even function with some of these factors during the backward pass.
I mean I am reassured that it is symmetric, but this is cold comfort when it works so poorly.

...Here is what I am going to do tomorrow.

I need to make sure that the updates actually converge to the covariance matrix for the big data examples. I haven't checked that yet.

I'll bring back the original update and test it again.

If it turns out that it does converge then what I am going to have to look into is somehow regularizing these updates so that they do sensible things.

I am a bit annoyed at this right now. Maybe I really should just muster the will and bring in the cuSolver functions. I will give this a day or two more and then move in that direction.

This is taking way longer than I thought. Considering I made up my mind to implement PRONG at the start of the month I have hard time understanding why is it still not done.