I spent the last hour since I got up crafting that reply, but now it is time that I start this.
I want to do more testing today. I am very suspicious of the inverse update and as a sanity check, I want to see that the regular update is symmetric and that it works properly for all learning rates.

Even if I set something like 0.1, none of the updates should blow up. This is something that should have occured to me yesterday.

If I cannot get these iterative updates to work, I will spend the next week putting in the cuSolver functions.