I am dead. I have no choice at this point. I'll put in the cuSolver module and start if off with the portf function which is the Cholesky decomposition.
Before that I'll study the decomposition algorithm so I understand what it is doing. Maybe I'll be fortunate and it will turn out that the Cholesky decomposition will be rather inexpesive. It is certainly costly now in the kernel I made. I really suspecting the NVCC compiler did not do a good job optimizing it, but that does not matter anymore.

I'll leave the code at this point in time in a separate branch just in case I need it later.