Those posts are incredibly time and energy consuming for me to make. Let me take the time to talk about the idea I had last night.
Originally, the great gust of insight I had a week ago on how to remove the reprojection steps is quite simple.

Since the updates to the whitening matrix must satisfy `((x - c) U W + b) R = ((x - c') U' W' + b') R'` in order to achieve that, the weight matrix `W` must be reprojected after every update of the whitening matrices so that the functionality of the network does not change.

The idea I had was a very simple one. Rather than do an explicit reprojection of W which would be inefficient, it would be better to use a matrix `M` instead which is equal to `U W R`. During the forward pass, I just use `M` as usual, but on the backward pass, I'd multiply the gradients coming from the top by `R` and the whiten the inputs by `W`.

This is actually a pretty brilliant idea and would give me the updates for `W`. Unfortunately, what occured to me tonight is that the update to `W` is not the same as the update to `M`. It is possible to show this on a simple scalar example.

`x * u * w * r * z`

Suppose the error `z` is 1 and the input `x` is 1 and the regular 'matrix' `w` is also 1. The whitening matrices `u` and `r` are 2 and the goal is to update `w` in `m = u * w * r = 2 * 1 * 2 = 4`.

`d w / d (x * u * w * r * z) = x * u * r * z`

Plugging in the numbers: `1 * 2 * 2 * 1 = 4`. The gradient for `w` is therefore 4. Suppose the `m` is updated.

`m = u * w * r = 2 * 4 * 2 = 16`

The new `m` is 16 past the update.

However if the `m = 4` is updated directly that would just result in `m = 8` which is incorrect and which I've been trying to do with the whitening method yesterday to no success. Even if the inverse Cholesky update worked, I would not have gotten the results I wanted.

The update for new `W` is `W' = U^T X Z R`, but what is really needed is an update for `M`. Therefore,

`U W' R = M' = U U^T X Z R R`

`U` and `R` are both symmetric Cholesky factors so the transpose in `U^T` does not matter. They are also inverse square roots of the covariance matrix. `U U` and `R R` and therefore just the inverses of the covariance matrix - also known as the information matrix.

So I do not need the inverse Cholesky factors, but rather their squares. Instead of butting my head trying to get the inverse update formula to work what I should do instead is implement Sherman-Morrison for starters. That should help stabilize things since I suspect that the Cholesky updates are greatly numerical unstable as written anyway.

```
A^-1 = U U
B^-1 = R R
M' = A^-1 (X Z) B^-1
```

Note that what I've eventually derived here is just the standard block diagonal K-FAC update except through a completely different path.

In the K-FAC the update is `A^-1 G B^-1` where `G` is the gradient. But rather than reproject the gradients which could be as large as the number of weights, what I could do is `(A^-1 X) (Z B^-1)` instead which would in general be a lot more efficient.

Not to mention what I have here now is a general scheme for how to do whitening as a part of the backwards pass. K-FAC method operates in the optimization pass instead. This means I could use this in a recurrent net and in nets with differentiable plasticity where I could not reproject the weight matrix even if wanted to do it.

This might be worth publishing a paper on.

Phew. Let me get breakfast. With this insight I am certainly going to be done with this by the end of the month.