I wish I could say I would resume programming immediately, but it seems I will have other obligations to take care off that will keep me away from the screen.
https://arxiv.org/pdf/1807.04912.pdf
Perceptrons from Memristors

Besides, I want to do some reading and studying. I want to read the above paper, watch the lecture on numerical stability of algorithms and reread the K-FAC papers.

I am not sure I'll manage all of that today. It will probably go on into tomorrow.

In order to do the inverses properly for batch learning I am going to have to put in Cholesky decomposition from the cuSolver library, no way around it, but it is fine. Cholesky in particular works on symmetric matrices so it does not conflict with row major standard that I want to impose in Spiral.

Apart from that I am actually going to have to ask on the ML sub for somebody to explain me the reasoning for the tridiagonal approximations. There are still major parts in the K-FAC paper that I still do not understand.

I have to go for now.