"5:45pm. What a waste of time today was. Sometimes I just feel like doing random reading while I think. Let me take a look at the conference notes and then I will call it a day.
6:20pm. 31/55. https://arxiv.org/abs/1802.07239
Continual Reinforcement Learning with Complex Synapses

This seems like it could be interesting. There is a bunch of useful stuff in the conference it seems.

6:25pm.  https://arxiv.org/abs/1706.03100
Decoupling Gradient Like Learning Rules from Representations

"This paper: generalized Amari’s gradient like-learning rule [2] to naturalized learning rule, including
TD-like-algorithms, policy gradient algorithms, and accelerated gradient methods."

Oh, how did I miss this?

I am going to have to read this one.

6:55pm. Done with lunch. I'll call it a day here. I really want to read the above paper, but at this point my energy is well and truly gone. I wasted quite a lot of it today. Hopefully tomorrow I will be able to more motivated learning. Besides the paper I really want to understand the reasoning behind the tridiagonal approximation. I still do not understand the recurrent K-FAC paper and I haven't even seen the convolutional K-FAC one.

I feel like it is time to start brining the natural gradient arc to a close. And the best way to do that is to start it properly in the first place."