"11am. Done with the first paper. I am not sure what to make of it. Let me focus on the second for a while. I am having real trouble focusing.
11:40am. The `Decoupling Gradient Like Learning Rules from Representations` paper was nothing. It is just a bunch of theory filler.

I could have skipped the two papers. The first paper is interesting in a sense, but I am just not that familiar with ODEs. There is probably a better way to do long term memories that meshes with the general DL flavor. Maybe RNNs with memory, but no BPTT? Something like that. This is why I am aiming for differentiable plasticity.

For the rest of the day and maybe even tomorrow, I will really focus on the K-FAC papers. I really want to figure out how the tridiagonal approximation works. I might have to ask on the ML sub for this.

12:20pm. Looking at page 34/58 of the K-FAC paper, it seems Martens was aware of the K-FAC projection.

Let me move to the recurrent paper. I still do not understand the tridiagonal approximation. I am going to have to figure out what the conditional covariance is as well.

12:45pm. Ugh, I am going to have to ask about the tridiagonal approximation. And the recurrent paper makes no sense to me.

A Kronecker-factored approximate Fisher matrix for convolution layers
https://arxiv.org/abs/1602.01407

Let me read this and then I will do it. After I am done with this and asking I'll take a break since I spent my breakfast reading..

Scalable Trust-Region Method for Deep Reinforcement Learning Using Kronecker-Factored Approximation
https://www.youtube.com/watch?v=0rrffaYuUi4

I've never watched this video before.

http://proceedings.mlr.press/v37/grosse15.html
Scaling up Natural Gradient by Sparsely Factorizing the Inverse Fisher Matrix

Never read this paper before either...ah, that is the RBM paper. I'll skip it.

1:50pm. Let me make the question and then I will watch the K-FAC for RL video. My mind feels kind of dull and I feel like I am not learning from my effort here.

2:05pm. "I have a decent grasp of the block diagonal K-FAC update, and I even understand how to derive it from the PRONG update, but the tri-diagonal K-FAC update is beyond my understanding.

Why are graphical models used in the context of NN in order to derive the update? What is the dependent covariance? What are the exact steps in the paper taken to derive the updates? What would the explicit whitening in the style of PRONG look like as a part of a NN architecture for the tri-diagonal approximation? There are probably more question after this, but I will have to get through these until I can find them."

2:10pm. https://www.reddit.com/r/MachineLearning/comments/8zkjwb/d_is_there_a_simpler_explanation_of_the_kfac

Here is the question.

Let me take a break and then I will watch the video. After that I will seriously start work on bringing in the Cholesky decomposition. For the next few days I'll be working only on putting in more Cuda library functions."