"3:25pm. That was a long break. Let me watch the video. I was thinking of how I would deal with regularization in the Woodbury identity and Sherman Morrison. There simply just isn't any way. If the denominator term is non-invertible the only choice that I have is to just throw it away and that is it.
Maybe eventually somebody will figure out a better way to do all this whitening stuff, and that would include making it fit for recurrent nets. I am not sure that the current whitening schemes will be suitable for that. I intend to backward chain the updates, but maybe that won't really fit.

https://www.youtube.com/watch?v=0rrffaYuUi4

Well, let me watch this video and maybe I'll call it a day. Tomorrow I will really roll up my sleeves and just code all the way until I am done with this. Then I will be able to start work on Zap Q.

Hopefully after I have that I will have enough of a starter kit for the first poker player.

4:40pm. https://youtu.be/0rrffaYuUi4?t=2017

"In RL we care more about sample efficiency than wall clock time."

Amen.

5:05pm. https://youtu.be/0rrffaYuUi4?t=3303

"Do you have an idea how this compares to PPO?"

Good question. I actually do not think this was in the ACKTR paper.

5:30pm. https://www.reddit.com/r/reinforcementlearning/comments/8zm24l/kfac_for_rl_acktr_performance_figures/

Made a thread on the RL sub. The performance figures of ACKTR are rather new to me and they might be of interest to others.

5:35pm. https://arxiv.org/pdf/1712.03112.pdf
Effective Extensible Programming: Unleashing Julia on GPUs

I got this paper from Mason Protter after asking him whether Julia supports GPU garbage collection. It seems it does. Actually being able to do something like this makes it quite a lot more attractive as an ecosystem for implementing a Spiral variant as a DSL in it.

Let me read the paper and I will call it a day.

This is how it usually goes to me. It feels like I spent most of my day meditating. I certainly did not spent the entirety of it in intensive learning though I certainly tried. The main principle to which I try to hold myself is not to play games or read manga during the working part of the day, so my mind can focus on my goals. I wish I could spend my time better, but until I crack the puzzle and tap into the essence of creativity more directly, it will have to be like this."