"7:50am. I am up. Let me explain the trick to that guy.
```
I'll just show you directly, quick and dirty. First of all, consider the equality constraint from the BPRONG paper that the method most satisfy after the update to the whitening params for all `x`. I am getting rid of the centering part from the paper.

`z = (x U W + b) R = (x U' W' + b') R'`

In order to make the two sides equal each other we set.

`U W R = U' W' R'` and `b R = b' R'`.

Now imagine instead of having `U`, `W` and `R` separately that we have a single matrices `M = U W R = U' W' R'` and `v = b R = b' R'` that we intend to update implicitly. Imagine that we are also tracking the inverse square covariance matrices `U` and `R`.

Here is how the gradient for `W` would be done.

`dW = (x U)^T (dz R) = U^T x^T dz R`

This is just the backward step for the matrix multiplication. But note that this is the gradient for `W` and not `M`. The step to make it an update for `M` is this.

`dM = U dW R = U U^T x^T dz R R`.

At this point it is assumed that `U` and `R` are symmetric. With that `U U^T` actually becomes the inverse of the covariance matrix.

`(U U^T) (x^T dz) (R R)`

When written like the above, the expression exactly the same as the block diagonal K-FAC update.

`((U U^T) x^T) (dz (R R))`

However writing it like this would be much more efficient with small batch sizes. This is essentially the PRONG/K-FAC update. One other difference is that in this update you would do this kind of thing during the backward step while K-FAC operates during the optimization pass if I've read the paper correctly. The beauty of this thing is that it should be possible to use it when the weight matrix `W` is not just a matrix of weights, but like in differentiable plasticity also depends on the data and the previous time steps where reprojection would actually be impossible. With this it also becomes unnecessary to track the inverse Cholesky factors, instead their squares are needed which is convenient because the inverse rank one Cholesky update from the CMA paper is quite numerically unstable.

The update for `v` is similar hopefully should be obvious.

If you are using TF or PyTorch you might be able to test this even earlier than me. Since I am doing this bare bones I need to set a bunch of infrastructure in place to get this to work first.
```

8:30am. theophrastzunz:

```
Ok, sorry for the delay. As I recall the derivation in the K-FAC paper is motivated by the empirical observation that the inverse FIM is block-diagonally dominant. If you look at Martens and Gross 2015, figure 6, you'll see this in the bottom right corner. It's not very principled, since ideally to motivate this you'd want to show that the inverse FIM is block-diagonally dominant wrt some norm (l_2,l_1, Frobenius).

I misspoke eariler. So it's not a block-diagonal approximation to the FIM but rather the inverse FIM. In this sense, they seek a precision matrix that closely matches the the inverse (FIM-1). You wouldn't necessarily get this from a block covariance since the inverse of a sparse, block structured matrix isn't necessarily sparse nor does it have the same block structure. The exception to this is the diagonal case, where each of the blocks acts on a different subspace.

PRONG in it's basic form only applies whitening to the activations of each layer. Therefore it cannot capture the inter-layer dependence. In a more general sense, mirror descent on which prong is based is entirely compatible with Natural gradient descent, so yes, you could generalize PRONG appropriately or rather see natural gradient descent as mirror descent.

I don't follow what you mean here

When I rederived the K-FAC update from PRONG, I made a significant gain because unlike in the K-FAC, the particular update operates not by projecting the weight gradients (though it could,) but by projecting the output gradients and the inputs which in online learning would be vastly more efficient.

Concerning this

It might also work as is for recurrent nets, including those like differentiable plasticity which have adaptive weight matrices that are impossible to explicitly reproject because they depend on data.

FIM most decidedly depends on the data. In natural gradient descent you don't project the weights, but project the gradients. It is generalizable to RNNs, and there have been some papers this year (but I don't have them on hand & haven't read them yet).
```

This wasn't what I was asking, but nonetheless this is interesting. I didn't really pay attention to this in the paper, but maybe I should. Let me take a look again.

8:50am. "> Ok, sorry for the delay. As I recall the derivation in the K-FAC paper is motivated by the empirical observation that the inverse FIM is block-diagonally dominant. If you look at Martens and Gross 2015, figure 6, you'll see this in the bottom right corner. It's not very principled, since ideally to motivate this you'd want to show that the inverse FIM is block-diagonally dominant wrt some norm (l_2,l_1, Frobenius).

Ah, I see. I wasn't really paying attention to this concern in the paper, but it makes sense to me now.

> You wouldn't necessarily get this from a block covariance since the inverse of a sparse, block structured matrix isn't necessarily sparse nor does it have the same block structure. The exception to this is the diagonal case, where each of the blocks acts on a different subspace.

Yeah, I missed this. Still, this leaves the mystery the tridiagonal update intact. I'll read [the paper](https://arxiv.org/abs/1303.5237) when I am done with this post.

> I don't follow what you mean here

I explain the trick I discovered a few days ago, but I've yet to test [here](https://www.reddit.com/r/MachineLearning/comments/8zfxmy/d_activation_function_that_preserves_mean/e2l83yk/?st=jjqrfaxp&sh=4ca0f768)."

8:50am. Let me read the Kalman smoothing paper now."