"12:50pm. Ah, forget anime. I had enough of resting.
Let me read that paper and then I will work on putting in the Cholesky decomposition into the library.

2:35pm. bbsome:

```
Not familiar with PRONG, but essentially one can show that finding the MAP of a Linear Gaussian System with pre-specified conditions is equivalent to a least-squares problem where one of the matrices is block-tri diagonal (for a linear Gaussian system the precision matrix is tri-diagonal of the concatenated state). Hence there is an equivalence between message passing for finding the MAP in GM and inverting a tri-diagonal matrix. In fact, if you derive the equation for inverting a tri-diagonal system you will see that you will get the same equations. For reference: https://arxiv.org/pdf/1303.5237.pdf
```

```
Thanks for the reference. The paper is quite hard to read though, and I can't follow it at my level of understanding. Do you have any particular suggestions for the path I should follow in order to get an understanding of this? If not I'll just start off by studying Kalman smoothing and follow the references in the paper you've linked me to.

Also it seems there is a bunch of material on inverting a block tridiagonal matrix on Google, so maybe that will be enough to give me an understanding of this.
```

With this I am truly done with reading for the day. Unless somebody explains to me the update in more depth, I will leave it alone for the time being until I am done implementing the iterative PRONG.

For now let me start programming. I've been under such strain with all this studying, interacting and the failed attemtps to get the inverse Cholesky update operational and I want to do some slow and concrete work. Adding Cuda library functions is tedious and time consuming so I won't force myself too much here. Let me go at it one at a time starting with the cuSolver dependency."