Done with that as well. I think I am going to call it a day here.
100 lines of code always take a while regardless of the task. With this I have the Cholesky decomposition and will be able to use this to invert covariance matrices should I need to. Thankfully I already have the triangular solve in the library.

This is just the backup plan in case the Woodbury identity turns out to be unstable as an update.

Tomorrow...hmmm...I need to think of what I am going to do. Ahhh, yes the allocator. That will be my number 1 thing on the list. I need to have it dump empty pointers at the top of the stack. This is how I did it in the previous allocator. I pretty much forgot how it works by now and will have to refresh my memory. I've been meaning to do this for a while anyway.

But after that...I am thinking of adding matrix-vector multiplications. I'll need them for Sherman-Morrison. I am thinking of forgetting batch learning and simply trying out online learning for a while. I know that `gemv` for example is 2-3 times faster than `gemm`. Doing it 100% online will allow me to use packed symmetric representations.

Also at the present time, I actually do not know how to use Zap Q in batch mode. I asked the author, but he seems to be in a hush hush situation. I am not sure where he is interning, but he cannot share much information at the moment.

I'll also need to add the symmetric rank-one update function for Sherman-Morrison. That one will be a bit challening as I do not want to pass alpha into it on host. I might need to change `CudaBlas` pointer modes just for this. I'll probably make a special version of it just for this.

Well, I'll get through it somehow. I just need to sit myself down like today and crack it out. I'll be done before I know it. I am actually not that far from completing PRONG. Once I am done with that and Zap Q I might have something substantial on my hands like a real power.

Spiral is an accomplishment? No, I cannot accept that. I want something that I can use to affect the real world. Using Spiral to complete these algorithms will be a real measure of success.