That was a nice long nap. Let me start work on the allocator.
That I do not completely understand K-FAC despite studying natural gradients for so long is absurd. It bothers me and I vow to figure this out by the end of the month. I also need to figure out how the recurrent version works as well.

After I have that understanding, I'll be able to move on from this to bigger and better things. It is easy to fall into a spell where you spend months and months going nowhere. I know that very well.

I do not want to spend too long in the 'make the agent' stage.