"4:55pm. https://youtu.be/4vGiHC35j9s?t=514 "Dot products measure similarity."
Well I knew this one, but forgot it along the way. It is interesting to consider in the context of covariances.

5:10pm. "If you are serious about machine learning I strongly suggest that you read chapter 4 for the course of Kevin Murphy's textbook."

6:25pm. https://youtu.be/4vGiHC35j9s?t=4418

These updates remind me a bit of the Kalman filter. The video is almost over.

7pm. Done with lunch. Let me finally finish the damn video.

7:10pm. It feels like a lid has been lifted on me. I am really much more comfortable with this Bayesian stuff than I was before. Previously I would have recoiled at all of this, but as NG methods are the real metalearning methods I am greatly motivated to understand them. I finally have an actual reason to study graphical models.

Change of plans. I am going to give Bayesian ML another try starting with Nando's course. Forget adding more Cuda functions to the library. That can wait until I am good and ready.

I understand NNs and I understand RL, but this is still a mystery to me. To be honest, I thought it was completely useless up to now, but if it can be used to do metalearning in the style of K-FAC then I will be there. I think the tridiagonal K-FAC update essentially models the gradients using GPs.

I am definitely going to be looking at the theorems for GPs in chapter 4 of Kevin Murphy's book."