Studying Bayesian ML is as about as enjoyable as reading other people's source code. Ugh.
It is not going as planned. I spent most of the day studying, but I do not think I learned anything. Towards the end yesterday I had a blast of insight with respect to Gaussian processes, but that quickly became murky.

Honestly, it is not looking like I will be able to figure this out through my present approach.

Well, maybe I'll look into what bbsome said, something about Gaussian elimination for the block trigonal matrix.

But forget that for a bit. Let me follow through with my plan for as much as I can.

https://www.youtube.com/playlist?list=PLuRaSnb3n4kSQFyt8VBldsQ9pO9Xtu8rY

I'll take a look at the lectures from this course. Maybe they will be more digestible.