...I do not think I'll able to figure out the tridiagonal update, but if it merely how to whiten multiple layers then that won't be a problem. I've figured it out.
All I need to do is concatenate inputs, do the projection using the precision matrix and then split them up. This could all be done during the backward pass. There would be some engineering challenges in grouping the layers like that, but they could be done.

I won't worry about the tridiagonal update anymore.

Now...I am a bit more concerned about recurrent nets. I also do not not understand the recurrent K-FAC update either.

I need something for that too, but I think that the raw whitening of the kind I will be doing is not the right answer here.

The right answer would be more along the lines of what Zap Q is doing, but that algorithm is basically a complete black box.

Nonetheless, the individual steps are surprisingly sensible. Maybe I could take advantage of that? It might be worth to gamble a little and check out whether it would do better variance reduction than the whitening procedure.

Well, it is not my still to plunge into the dark completely like that. With PRONG and the recent multi layer whitening idea, I know what my basis for doing that is.

I think I'll just wait instead. I am not in a hurry so unless the field enters a long dry spell I'll have plenty of things to do anyway.

Seriously fuck math. It is like an endless jungle of it. This is not the kind of work suitable for my peanut brain. I need a better one.

https://www.youtube.com/watch?v=ZA61mcMl63k&list=PLuRaSnb3n4kSQFyt8VBldsQ9pO9Xtu8rY&index=2

Let me watch this lecture and then I will call it a day. I'll watch third of the course tomorrow and the rest in the days after that.