I am up. It is time to continue that course.
Also I am continuing to work on that idea that I had. I think I know what sort of architecture would be necessary to implement it from a whitening perspective, but it would be quite a beast. It is would be far to better to implement it in the abstract. Well, I do not feel like going into it right now.

Right now I am still a bit groggy. Let me just watch the next lecture.

I like how the lecturer says that NNs are easy so we should just learn the hard stuff instead.