"9:30am. Looking at the formula more closely is making me realize what an idiot I was. I've been thinking about the denominator so much, but made the mistake of thinking of that operator as a `-` instead of a `+`. `1 + v^t A^-1 u` will always be >= 1 with positive semidefinite matrices.
That will make things much, much easier on me since I have guaranteed stability in the algorithm. I do not have to add noise to the inputs, I can just expect this to work.

With this one of the main obstacles that I've been worrying about has been cleared and so simply.

The Kalman matrix gain in Zap Q on the other hand...

I think A rather than A^-1 in that algorithm is supposed to have negative eigenvalues.

9:40am. Oh, so a matrix is invertible only if it does not have a zero eigenvalue.

"It is assumed throughout this section that A is Hurwitz: the real part of each eigenvalue is negative."

This is from the Zap Q paper. So should I initialize it to a negative identity?

9:45am. No forget it. I'll think about Zap Q when I come to it. Let me just focus on the task at hand. My mission for the day will be to get the inverse update to work."