"11:10am. https://github.com/amarack/python-rl/blob/master/pyrl/misc/matrix.py
This is really interesting. So it is possible to regularize the algorithm like this?

11:15am. No it would not prevent it from blowing up.

```
[|
    [|-9.815938; -1.877632|]
|]

[|
    [|97.35264; 18.43072|]
    [|18.43072; 4.525503|]
|]

---
-----
[|
    [|1.955139; 0.1827028|]
    [|0.1827028; 1.034948|]
|]

[|
    [|193.7053; 36.86144|]
    [|36.86144; 8.051005|]
|]
```

At any rate here is what is I am inputting into the algorithm. The thing at the bottom is supposed to be `C * C^-1`. The two updates are not inverses of each other at all. What the hell? I set both alpha and beta to 1, so it cannot be because of the learning rate. Did I make a mistake in the implementation?

11:20am. I have no idea. I guess I'll compare it with the Python implementation that I found on the net.

11:25am. Let me commit here. I'll resume after breakfast."