I hadn't realized that going from 2d to 1d could be done that easily.
Also, I want to use the symmetric functions, but I cannot do the update in the last step right now.

Let me just get rid of the temporaries. This will be easy and it is something I absolutely must do. Making kernels that fuse all of this can happen later.