Ok, that works. Time for the next thing.
Let me give it a pass through the whole thing. I want to see if it blows up or if it does the sensible thing. I am still sore from the Cholesky update thing, but I should do more thorough testing before I move to implementing the layer.

I do not feel like doing that right now anyway, it is getting late and I do not have enough time to do anything significant for the rest of the day, so I might as well fill some holes for the next hour or so.