Is the router still messing up? Enough dallying, let me just do this thing. There just isn't that much to think about here. I need to implement PRONG now. I'll think about the rest later.
The great irony of the past few years is that with the right goal and knowledge, somebody could do what I did in 10-100x less time.

But that also means that once I grab hold of the thread, I will be able to make really rapid progress. Let me do PRONG and Zap Q. I can't be stuck on this forever. There are not that many improvements before going from the current trash tier algorithms to optimal ones and this stuff should be enough to get my foot through the door.

What I am doing here might be absolutely essentially, but it is also basic and not that important in the grand scheme of things. My job is definitely not to figure out these low level details of optimizing NNs, but to big picture thinking. With PRONG, Zap and Rudder in mind, I'll slowly but surely get to that point.

I think Rudder is one of the most important papers in the deep ML era, but Zap Q seems to be doing something to bridge whitening methods like PRONG and reward decomposition. I do not think that using explicit reward composition like Rudder does is something that will be done in the long term, but rather it will be used as a basis to find the true algorithm.

Zap Q might already be it, but if that is true then it should be used to replace PRONG in the internal layers rather than just as the critic.

To put it simply, all these methods need to be melted down and reforged into one. I think once that is done, the genie will be out of the bottle for good and the Ascent To The Transcendence will begin. Right now the world is still in the pre-Ascent period.

The fact that I changed my mind 3 weeks ago is that the 1000x improvement is already here is absolutely remarkable. Two months ago I was not sure whether that would even happen though I was hoping it would.

This boggles the imagination. These are the kinds of events that can only happen in the pre-Singularity period.